test[w2.5]: refactor tests for advanced search
