+++ HealthcareSystem/PatientWebAppUnitTests/CreateObjectsForTests/CreateExaminationSearchDTO.cs
@@ -15,7 +15,7 @@ namespace PatientWebAppTests.CreateObjectsForTests
 
         public ExaminationSearchDTO CreateInvalidTestObject()
         {
-            return new ExaminationSearchDTO(jmbg: "1309998775018", startDate: new DateTime(2020, 10, 10), endOperator: LogicalOperator.And, endDate: DateTime.Now, doctorOperator: LogicalOperator.And, doctorSurname: null, anamnesisOperator: LogicalOperator.And, anamnesis: "gdsgsdg");
+            return new ExaminationSearchDTO(jmbg: null, startDate: new DateTime(2020, 10, 10), endOperator: LogicalOperator.And, endDate: DateTime.Now, doctorOperator: LogicalOperator.And, doctorSurname: null, anamnesisOperator: LogicalOperator.And, anamnesis: "gdsgsdg");
         }
     }
 }
