+++ HealthcareSystem/PatientWebAppUnitTests/UnitTests/ExaminationControllerTests.cs
@@ -62,175 +62,27 @@ namespace PatientWebAppTests.UnitTests
             Assert.True(result is NotFoundObjectResult);
         }
 
-
         [Fact]
-        public void Advanced_search_examination_operator_and_anamnesis()
+        public void Advanced_search_examination()
         {
-            ExaminationService examinationService = new ExaminationService(_stubRepository.CreateExaminationStubRepository());
+            ExaminationController examinationController = SetupExaminationController();
 
             var examinationSearchDTOValidObject = _objectFactory.GetExaminationSearchDTO().CreateValidTestObject();
-            var result = examinationService.AdvancedSearch(examinationSearchDTOValidObject);
+            var result = examinationController.AdvanceSearchExaminations(examinationSearchDTOValidObject);
 
-            Assert.NotEmpty(result);
-        }
-
-        [Fact]
-        public void Advanced_search_examination_operator_or_anamnesis()
-        {
-            ExaminationService examinationService = new ExaminationService(_stubRepository.CreateExaminationStubRepository());
-            
-            var result = examinationService.AdvancedSearch(new ExaminationSearchDTO(jmbg: "1309998775018", startDate: new DateTime(2020, 10, 10), endOperator: LogicalOperator.Or, endDate: new DateTime(2020, 12, 12), doctorOperator: LogicalOperator.And, doctorSurname: null, anamnesisOperator: LogicalOperator.Or, anamnesis: "lklklkl"));
-
-            Assert.NotEmpty(result);
-        }
-
-        [Fact]
-        public void Advanced_search_examination_operator_and_surname()
-        {
-            ExaminationService examinationService = new ExaminationService(_stubRepository.CreateExaminationStubRepository());
-
-            var result = examinationService.AdvancedSearch(new ExaminationSearchDTO(jmbg: "1309998775018", startDate: new DateTime(2020, 10, 10), endOperator: LogicalOperator.Or, endDate: new DateTime(2020, 12, 12), doctorOperator: LogicalOperator.And, doctorSurname: "Mar", anamnesisOperator: LogicalOperator.And, anamnesis: null));
-
-            Assert.NotEmpty(result);
-        }
-
-        [Fact]
-        public void Advanced_search_examination_operator_or_surname()
-        {
-            ExaminationService examinationService = new ExaminationService(_stubRepository.CreateExaminationStubRepository());
-
-            var result = examinationService.AdvancedSearch(new ExaminationSearchDTO(jmbg: "1309998775018", startDate: new DateTime(2020, 10, 10), endOperator: LogicalOperator.Or, endDate: new DateTime(2020, 12, 12), doctorOperator: LogicalOperator.Or, doctorSurname: "gdgdg", anamnesisOperator: LogicalOperator.And, anamnesis: null));
-
-            Assert.NotEmpty(result);
-        }
-
-        [Fact]
-        public void Advanced_search_examination_start_date()
-        {
-            ExaminationService examinationService = new ExaminationService(_stubRepository.CreateExaminationStubRepository());
-
-            var result = examinationService.AdvancedSearch(new ExaminationSearchDTO(jmbg: "1309998775018", startDate: new DateTime(2020, 10, 10), endOperator: LogicalOperator.Or, endDate: new DateTime(2020, 12, 12), doctorOperator: LogicalOperator.And, doctorSurname: null, anamnesisOperator: LogicalOperator.And, anamnesis: null));
-
-            Assert.NotEmpty(result);
-        }
-
-        [Fact]
-        public void Advanced_search_examination_operator_or_end_date()
-        {
-            ExaminationService examinationService = new ExaminationService(_stubRepository.CreateExaminationStubRepository());
-
-            var result = examinationService.AdvancedSearch(new ExaminationSearchDTO(jmbg: "1309998775018", startDate: DateTime.Now, endOperator: LogicalOperator.Or, endDate: new DateTime(2020, 12, 12), doctorOperator: LogicalOperator.And, doctorSurname: null, anamnesisOperator: LogicalOperator.And, anamnesis: null));
-
-            Assert.NotEmpty(result);
-        }
-
-        [Fact]
-        public void Advanced_search_examination_operator_and_end_date()
-        {
-            ExaminationService examinationService = new ExaminationService(_stubRepository.CreateExaminationStubRepository());
-
-            var result = examinationService.AdvancedSearch(new ExaminationSearchDTO(jmbg: "1309998775018", startDate: DateTime.Now, endOperator: LogicalOperator.And, endDate: new DateTime(2020, 12, 12), doctorOperator: LogicalOperator.And, doctorSurname: null, anamnesisOperator: LogicalOperator.And, anamnesis: null));
-
-            Assert.Empty(result);
+            Assert.True(result is OkObjectResult);
         }
 
         [Fact]
         public void Advanced_search_non_exitent_examination()
         {
-            ExaminationService examinationService = new ExaminationService(_stubRepository.CreateExaminationStubRepository());
+            ExaminationController examinationController = SetupExaminationController();
 
             var examinationSearchDTOInvalidObject = _objectFactory.GetExaminationSearchDTO().CreateInvalidTestObject();
-            var result = examinationService.AdvancedSearch(examinationSearchDTOInvalidObject);
+            var result = examinationController.AdvanceSearchExaminations(examinationSearchDTOInvalidObject);
 
-            Assert.Empty(result);
+            Assert.True(result is NotFoundObjectResult);
         }
 
-        [Fact]
-        public void Is_specification_anamnesis_true()
-        {
-            ExaminationAnamnesisSpecification anamnesisSpecification = new ExaminationAnamnesisSpecification("bol");
-            
-            var examinationValidObject = _objectFactory.GetExamination().CreateValidTestObject();
-            var result = anamnesisSpecification.IsSatisfiedBy(examinationValidObject);
-
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void Is_specification_anamnesis_false()
-        {
-            ExaminationAnamnesisSpecification anamnesisSpecification = new ExaminationAnamnesisSpecification("fdfd");
-
-            var examinationValidObject = _objectFactory.GetExamination().CreateValidTestObject();
-            var result = anamnesisSpecification.IsSatisfiedBy(examinationValidObject);
-
-            Assert.False(result);
-        }
-
-        [Fact]
-        public void Is_specification_doctor_surname_true()
-        {
-            ExaminationDoctorSurnameSpecification surnameSpecification = new ExaminationDoctorSurnameSpecification("markovic");
-
-            var examinationValidObject = _objectFactory.GetExamination().CreateValidTestObject();
-            var result = surnameSpecification.IsSatisfiedBy(examinationValidObject);
-
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void Is_specification_doctor_surname_false()
-        {
-            ExaminationDoctorSurnameSpecification surnameSpecification = new ExaminationDoctorSurnameSpecification("kllk");
-
-            var examinationValidObject = _objectFactory.GetExamination().CreateValidTestObject();
-            var result = surnameSpecification.IsSatisfiedBy(examinationValidObject);
-
-            Assert.False(result);
-        }
-
-        [Fact]
-        public void Is_specification_start_date_true()
-        {
-            ExaminationStartDateSpecification startSpecification = new ExaminationStartDateSpecification(new DateTime(2019, 1, 11));
-
-            var examinationValidObject = _objectFactory.GetExamination().CreateValidTestObject();
-            var result = startSpecification.IsSatisfiedBy(examinationValidObject);
-
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void Is_specification_start_date_false()
-        {
-            ExaminationStartDateSpecification startSpecification = new ExaminationStartDateSpecification(new DateTime(2020, 12, 12));
-
-            var examinationValidObject = _objectFactory.GetExamination().CreateValidTestObject();
-            var result = startSpecification.IsSatisfiedBy(examinationValidObject);
-
-            Assert.False(result);
-        }
-
-        [Fact]
-        public void Is_specification_end_date_true()
-        {
-            ExaminationEndDateSpecification endSpecification = new ExaminationEndDateSpecification(new DateTime(2020, 12, 12));
-
-            var examinationValidObject = _objectFactory.GetExamination().CreateValidTestObject();
-            var result = endSpecification.IsSatisfiedBy(examinationValidObject);
-
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void Is_specification_end_date_false()
-        {
-            ExaminationEndDateSpecification endSpecification = new ExaminationEndDateSpecification(new DateTime(2020, 10, 10));
-
-            var examinationValidObject = _objectFactory.GetExamination().CreateValidTestObject();
-            var result = endSpecification.IsSatisfiedBy(examinationValidObject);
-
-            Assert.False(result);
-        }
     }
 }
